+++ lib/runners/processor/checkstyle.rb
@@ -3,7 +3,7 @@ module Runners
     include Java
 
     Schema = StrongJSON.new do
-      let :runner_config, Schema::RunnerConfig.base.update_fields { |fields|
+      let :runner_config, Schema::BaseConfig.base.update_fields { |fields|
         fields.merge!({
                         config: string?,
                         dir: optional(enum(string, array(string))),
@@ -24,6 +24,8 @@ module Runners
         severity: string?,
       )
     end
+
+    register_config_schema(name: :checkstyle, schema: Schema.runner_config)
 
     def self.ci_config_section_name
       'checkstyle'