+++ lib/runners/processor/tyscan.rb
@@ -4,7 +4,7 @@ module Runners
 
     # Define schema in sider.yml
     Schema = StrongJSON.new do
-      let :runner_config, Schema::RunnerConfig.npm.update_fields { |fields|
+      let :runner_config, Schema::BaseConfig.npm.update_fields { |fields|
         fields.merge!({
                         config: string?,
                         tsconfig: string?,
@@ -17,6 +17,8 @@ module Runners
         message: string,
       )
     end
+
+    register_config_schema(name: :tyscan, schema: Schema.runner_config)
 
     DEFAULT_DEPS = DefaultDependencies.new(
       main: Dependency.new(name: "tyscan", version: "0.3.1"),