+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - [RuboCop] Use `--out=FILE` option [#769](https://github.com/sider/runners/pull/769)
 - [go-vet][Golint][Go Meta Linter] Add deprecated warning [#767](https://github.com/sider/runners/pull/767)
 - [Goodcheck] Update goodcheck requirement from 2.4.5 to 2.5.0 [#775](https://github.com/sider/runners/pull/775)
+- Add Runners::Config, which is responsible for sider.yml [#763](https://github.com/sider/runners/pull/763)
 
 ## 0.20.0